requirements.txt
streamlit
diffusers
transformers
torch
accelerate
Pillow
safetensors